Ádds database

+++ croins/Data/Model/DataInputIn.swift
@@ -2,11 +2,11 @@ import Foundation
 
 class DataInputIn: Codable {
     let title: String
-    let value: Decimal
+    let value: Double
     let date: Date
     //let isRecurrent: Bool
     
-    init(title: String, value: Decimal, date: Date)  {
+    init(title: String, value: Double, date: Date)  {
         self.title = title
         self.value = value
         self.date = date
